refactor: rename task update route to task index and remove unused import dialog
